Ensure Gauge secrets are ignored by source control
By convention we will allow Gauge secrets to be stored in a [Gauge
properties file][1] called `secrets.properties`.  Secrets must never be
stored in source control, so this commit ensures that any files with
that name are ignored by Git.

NB Our Gauge secrets don't have to be stored in a properties file - they
can also be [set as environment variables][2].

[1]: https://docs.gauge.org/configuration.html
[2]: https://docs.gauge.org/configuration.html#precedence-of-environments-when-running-gauge